+++ R/HDI.R
@@ -6,8 +6,8 @@
 #'              distribution that lies within a region of practical equivalence.
 #'
 #' @param x A vector of values from a probability distribution (e.g., posterior
-#'        probabilities from MCMC sampling), or a \code{stanreg}- or
-#'        \code{stanfit}-object.
+#'        probabilities from MCMC sampling), or a \code{stanreg},
+#'        \code{stanfit}, or \code{brmsfit} object.
 #' @param prob Scalar between 0 and 1, indicating the mass within the credible
 #'        interval that is to be estimated.
 #' @param rope Vector of length two, indicating the lower and upper limit of a
@@ -19,60 +19,69 @@
 #'        (for \code{rope()}) on the values of the posterior distribution, before
 #'        calculating the rope based on the boundaries given in \code{rope}. Note
 #'        that the values in \code{rope} are not transformed.
+#' @param type For mixed effects models, specify the type of effects that should
+#'        be returned. \code{type = "fixed"} returns fixed effects only,
+#'        \code{type = "random"} the random effects and \code{type = "all"} returns
+#'        both fixed and random effects.
 #'
 #'
 #' @return For \code{hdi()}, if \code{x} is a vector, returns a vector of length two
 #'         with the lower and upper limit of the HDI; if \code{x} is a
-#'         \code{stanreg}-object, returns a tibble with lower and upper HDI-limits
-#'         for each predictor. For \code{rope()}, returns the proportion of values
-#'         from \code{x} that are within the boundaries of \code{rope}.
+#'         \code{stanreg}, \code{stanfit} or \code{brmsfit} object, returns a
+#'         tibble with lower and upper HDI-limits for each predictor.
+#'         For \code{rope()}, returns the proportion of values from \code{x}
+#'         that are within the boundaries of \code{rope}.
 #'
 #' @details Computation for HDI is based on the code from Kruschke 2015, pp. 727f.
 #'
 #' @references Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
 #'
 #' @examples
-#' library(rstanarm)
-#' fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
-#' hdi(fit)
+#' if (require("rstanarm")) {
+#'   fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
+#'   hdi(fit)
 #'
-#' # fit logistic regression model
-#' fit <- stan_glm(
-#'   vs ~ wt + am,
-#'   data = mtcars,
-#'   family = binomial("logit"),
-#'   chains = 1
-#' )
-#' # compute hdi, transform on "odds ratio scale"
-#' hdi(fit, trans = exp)
+#'   # fit logistic regression model
+#'   fit <- stan_glm(
+#'     vs ~ wt + am,
+#'     data = mtcars,
+#'     family = binomial("logit"),
+#'     chains = 1
+#'   )
+#'   # compute hdi, transform on "odds ratio scale"
+#'   hdi(fit, trans = exp)
 #'
-#' # compute rope, on scale of linear predictor. finds proportion
-#' # of posterior distribution values between -1 and 1.
-#' rope(fit, rope = c(-1, 1))
+#'   # compute rope, on scale of linear predictor. finds proportion
+#'   # of posterior distribution values between -1 and 1.
+#'   rope(fit, rope = c(-1, 1))
 #'
-#' # compute rope, boundaries as "odds ratios". finds proportion of
-#' # posterior distribution values, which - after being exponentiated -
-#' # are between .8 and 1.25 (about -.22 and .22 on linear scale)
-#' rope(fit, rope = c(.8, 1.25), trans = exp)
+#'   # compute rope, boundaries as "odds ratios". finds proportion of
+#'   # posterior distribution values, which - after being exponentiated -
+#'   # are between .8 and 1.25 (about -.22 and .22 on linear scale)
+#'   rope(fit, rope = c(.8, 1.25), trans = exp)
+#' }
 #'
 #' @importFrom tibble as_tibble rownames_to_column
 #' @importFrom purrr map_dbl map_df
 #' @importFrom sjmisc rotate_df
 #' @export
-hdi <- function(x, prob = .9, trans = NULL) {
+hdi <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
   UseMethod("hdi")
 }
 
 
 #' @rdname hdi
 #' @export
-rope <- function(x, rope, trans = NULL) {
+rope <- function(x, rope, trans = NULL, type = c("fixed", "random", "all")) {
   UseMethod("rope")
 }
 
 
 #' @export
-hdi.stanreg <- function(x, prob = .9, trans = NULL) {
+hdi.stanreg <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
   # get posterior data
   dat <- x %>%
     tibble::as_tibble() %>%
@@ -82,12 +91,39 @@ hdi.stanreg <- function(x, prob = .9, trans = NULL) {
 
   colnames(dat) <- c("term", "hdi.low", "hdi.high")
 
-  dat
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
 }
 
 
 #' @export
-hdi.stanfit <- function(x, prob = .9, trans = NULL) {
+hdi.brmsfit <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
+  # check for pkg availability, else function might fail
+  if (!requireNamespace("brms", quietly = TRUE))
+    stop("Please install and load package `brms` first.")
+
+  # get posterior data
+  dat <- x %>%
+    tibble::as_tibble() %>%
+    purrr::map_df(~ hdi_helper(.x, prob, trans)) %>%
+    sjmisc::rotate_df() %>%
+    tibble::rownames_to_column()
+
+  colnames(dat) <- c("term", "hdi.low", "hdi.high")
+
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
+}
+
+
+#' @export
+hdi.stanfit <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
   # get posterior data
   dat <- x %>%
     as.data.frame() %>%
@@ -97,12 +133,13 @@ hdi.stanfit <- function(x, prob = .9, trans = NULL) {
 
   colnames(dat) <- c("term", "hdi.low", "hdi.high")
 
-  dat
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
 }
 
 
 #' @export
-hdi.default <- function(x, prob = .9, trans = NULL) {
+hdi.default <- function(x, prob = .9, trans = NULL, type = c("fixed", "random", "all")) {
   hdi_helper(x, prob, trans)
 }
 
@@ -128,28 +165,16 @@ hdi_helper <- function(x, prob, trans) {
 
 
 #' @export
-rope.default <- function(x, rope, trans = NULL) {
+rope.default <- function(x, rope, trans = NULL, type = c("fixed", "random", "all")) {
   rope_helper(x, rope, trans)
 }
 
 
 #' @export
-rope.stanreg <- function(x, rope, trans = NULL) {
-  # get posterior data
-  dat <- x %>%
-    tibble::as_tibble() %>%
-    purrr::map_df(~ rope_helper(.x, rope, trans)) %>%
-    sjmisc::rotate_df() %>%
-    tibble::rownames_to_column()
+rope.stanreg <- function(x, rope, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
 
-  colnames(dat) <- c("term", "rope")
-
-  dat
-}
-
-
-#' @export
-rope.stanfit <- function(x, rope, trans = NULL) {
   # get posterior data
   dat <- x %>%
     as.data.frame() %>%
@@ -159,7 +184,50 @@ rope.stanfit <- function(x, rope, trans = NULL) {
 
   colnames(dat) <- c("term", "rope")
 
-  dat
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
+}
+
+
+#' @export
+rope.brmsfit <- function(x, rope, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
+  # check for pkg availability, else function might fail
+  if (!requireNamespace("brms", quietly = TRUE))
+    stop("Please install and load package `brms` first.")
+
+  # get posterior data
+  dat <- x %>%
+    tibble::as_tibble() %>%
+    purrr::map_df(~ rope_helper(.x, rope, trans)) %>%
+    sjmisc::rotate_df() %>%
+    tibble::rownames_to_column()
+
+  colnames(dat) <- c("term", "rope")
+
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
+}
+
+
+#' @export
+rope.stanfit <- function(x, rope, trans = NULL, type = c("fixed", "random", "all")) {
+  # check arguments
+  type <- match.arg(type)
+
+  # get posterior data
+  dat <- x %>%
+    as.data.frame() %>%
+    purrr::map_df(~ rope_helper(.x, rope, trans)) %>%
+    sjmisc::rotate_df() %>%
+    tibble::rownames_to_column()
+
+  colnames(dat) <- c("term", "rope")
+
+  # check if we need to remove random or fixed effects
+  remove_effects_from_stan(dat, type)
 }